Major change to Function parameters. Added $ object as first parameter.
This allows developers to not need to cast the first argument to a $ or
get a new instance when a function is invoked.
- Also fixed an issue where a new instance of droidQuery would clear
event listeners on Views.
- Also modified documentation to work with new Function parameters, and
added some Methods to pass a context that can be used for creating the
droidQuery to pass to a Function, for different tasks (such as
$.Callbacks).