Modified build settings and included libraries.
This update is designed to make it easier to include droidQuery in a
project. Using the ant build sequence will allow droidQuery to build a
jar that includes all other jars. This is nice for projects that don't
use, for example, a different library containing NineOldAndroids. If
the developer chooses to use the jar that does not include this
dependency (and the CWAC-Task.jar dependency), he or she can add the
jars independently by looking in the libs/ and bin/ directory.